Handle SSLWantWriteError correctly
The Python SSL socket raises SSLWantWriteError if it is currently unable
to accept a write. The previous implementation treated this as an SSL
error and aborted the connection. This manifested as the connection
falling over during heavy upload activity.

This patch buffers the incompletely written data until the socket is
ready. It also stops reading from the pppd pty so that it doesn't act as
a buffer and add mountainous latency.

Upload under high load still seems to block for long periods (hundreds
of ms). Something is buffering - OpenSSL?