Send X-NE-SESSIONPROMPT header on login
This appears to fix a long-standing bug I have experienced where the ppp
link will come up, but no packets ever come out of the tunnel. This was
experienced with the official Linux client (at least as on v8.5.797) as
well as with nxBender. The condition will persist for a while, and then
randomly resolve (new sessions start working). Restarting the session
does not help.

Eventually I had a Windows machine handy and tried that during a problem
period, which worked. Sniffing the link showed this X-NE-SESSIONPROMPT
header being sent to the userLogin call; the Linux client does not do
this.

Adding this to the userLogin call seems to fix the issue.